engine-types: move enumLookup deeper in the stack
Although my reason for doing so has now disappeared.
